separates Loading Screen into multiple components
