+++ src/views/load/loadView.jsx
@@ -1,37 +1,8 @@
 import React from 'react';
-import { MuiThemeProvider, createMuiTheme } from '@material-ui/core/styles';
-import RouteButton from '../buttons/routeButton';
-
-let theme = createMuiTheme({
-  typography: {
-    useNextVariants: true,
-  },
-  palette: {
-    primary: {
-      main: '#00e676'
-    },
-  },
-});
+import LoadTitle from './loadTitle';
+import StartButton from '../buttons/startButton';
 
 let styles = {
-  initializeButton: {
-    width: '400px',
-    height: '300px',
-    flex: 'flex-grow',
-    flexBasis: '40%',
-    fontSize: '3em',
-    fontFamily: 'Courier',
-  },
-  screenOutterWrapper: {
-    height: '50%',
-    textAlign: 'center',
-    marginTop: '100px',
-  },
-  screenInnerWrapper: {
-    height: '50%',
-    fontSize: '8em',
-    textShadow: '2px 2px #00e676',
-  },
   buttonWrapper: {
     width: '100%',
     display: 'flex',
@@ -39,10 +10,6 @@ let styles = {
     bottom: '0',
     position: 'absolute',
     marginBottom: '70px',
-  },
-  descriptionWrapper: {
-    height: '50%',
-    fontSize: '3em',
   },
 }
 
@@ -52,63 +19,33 @@ export default class LoadView extends React.Component {
     let { savedGameFound, startNewGame, loadSavedGame } = this.props;
 
     return savedGameFound ? (
-      <div>
+      <React.Fragment>
+        <LoadTitle />
         <div
-          style={styles.screenOutterWrapper}
+          style={styles.buttonWrapper}
         >
-        <div
-          style={styles.screenInnerWrapper}
-        >
-          Green Lightning!
-        </div>
-        <div
-          style={styles.descriptionWrapper}
-        >
-          The galaxy's most action filled text based game
-        </div>
-        </div>
-        <div
-        style={styles.buttonWrapper}
-        >
-          <MuiThemeProvider theme={theme}>
-            <RouteButton
-              title={'New'}
-              routeTo={'/homescreen'}
-              onClick={startNewGame}
-              variant='contained'
-              color='primary'
-              style={styles.initializeButton}
-            >
-            </RouteButton>
-            <RouteButton
-              title={'Load'}
-              routeTo={'/homescreen'}
-              onClick={loadSavedGame}
-              variant='contained'
-              color='primary'
-              style={styles.initializeButton}
-            >
-            </RouteButton>
-          </MuiThemeProvider>
-        </div>
-      </div>
-    ) : (
-      <div>
-        Green Lightning!
-        <div />
-        The galaxy's most action filled text based game
-        <MuiThemeProvider theme={theme}>
-          <RouteButton
+          <StartButton
             title={'New'}
-            routeTo={'/homescreen'}
+            onClick={startNewGame}
+          />
+          <StartButton
+            title={'Load'}
             onClick={loadSavedGame}
-            variant='contained'
-            color='primary'
-            style={styles.initializeButton}
-          >
-          </RouteButton>
-        </MuiThemeProvider>
-      </div>
+          />
+        </div>
+      </React.Fragment>
+    ) : (
+      <React.Fragment>
+        <LoadTitle />
+        <div
+          style={styles.buttonWrapper}
+        >
+          <StartButton
+            title={'New'}
+            onClick={startNewGame}
+          />
+        </div>
+      </React.Fragment>
     );
   }
 }
